Separates rules that are exclusively used in internal experiments
