+++ src/main/resources/application.conf
@@ -1,6 +1,6 @@
 AsistEngine {
 // Override the default values here
-//    masterRulesPath = 
+      masterRulesPath = "/org/clulab/asist/grammars/master.yml"
 //    entityRulesPath =
 //     avoidRulesPath = 
 //       taxonomyPath =
