+++ src/main/resources/org/clulab/asist/grammars/entities.yml
@@ -3,24 +3,10 @@ vars: org/clulab/asist/grammars/vars.yml
 rules:
   - import: org/clulab/asist/grammars/ent_template.yml
     vars:
-      name: fire
-      priority: ${ rulepriority }
-      label: Fire
-      trigger: ${ fire_triggers }
-
-  - import: org/clulab/asist/grammars/ent_template.yml
-    vars:
       name: infrastructure
       priority: ${ rulepriority }
       label: Infrastructure
       trigger: ${ infrastructure_triggers }
-
-  - import: org/clulab/asist/grammars/ent_template.yml
-    vars:
-      name: foe
-      priority: ${ rulepriority }
-      label: Foe
-      trigger: ${ foe_triggers }
 
   - import: org/clulab/asist/grammars/ent_template.yml
     vars:
