+++ src/main/resources/org/clulab/asist/internal/internal_actions.yml
@@ -1,0 +1,18 @@
+vars: org/clulab/asist/grammars/vars.yml
+
+rules:
+  - import: org/clulab/asist/grammars/lvo_template.yml
+    vars:
+      name: craft
+      template_priority: ${ rulepriority }
+      label: Craft
+      trigger: ${ craft_triggers }
+      object_type: Concept?
+
+  - import: org/clulab/asist/grammars/lvo_template.yml
+    vars:
+      name: defeat
+      template_priority: ${ rulepriority }
+      label: Defeat
+      trigger: ${ attack_triggers }
+      object_type: Foe
