+++ src/main/resources/org/clulab/asist/internal/internal_entities.yml
@@ -1,0 +1,16 @@
+vars: org/clulab/asist/grammars/vars.yml
+
+rules:
+  - import: org/clulab/asist/grammars/ent_template.yml
+    vars:
+      name: fire
+      priority: ${ rulepriority }
+      label: Fire
+      trigger: ${ fire_triggers }
+
+  - import: org/clulab/asist/grammars/ent_template.yml
+    vars:
+      name: foe
+      priority: ${ rulepriority }
+      label: Foe
+      trigger: ${ foe_triggers }
