+++ src/main/resources/org/clulab/asist/internal/internal_master.yml
@@ -1,0 +1,68 @@
+taxonomy: org/clulab/asist/grammars/taxonomy.yml
+
+vars: org/clulab/asist/grammars/triggers.yml
+
+
+rules:
+
+  # Extract the entities
+  - import: org/clulab/asist/grammars/entities.yml
+    vars:
+      rulepriority: "1"
+
+  # Extract the internal-experiment only entities
+  - import: org/clulab/asist/internal/internal_entities.yml
+    vars:
+      rulepriority: "1"
+
+  # Extract the sentiments
+  - import: org/clulab/asist/grammars/sentiments.yml
+    vars:
+      rulepriority: "2"
+
+  # Extract the locations
+  - import: org/clulab/asist/grammars/locations.yml
+    vars:
+      first_priority: "2-3"
+      second_priority: "4+"
+
+  # Extract the base_concepts
+  - import: org/clulab/asist/grammars/base_concepts.yml
+    vars:
+      rulepriority: "3+"
+
+  # Extract time mentions, e.g., "30 seconds"
+  - import: org/clulab/asist/grammars/time.yml
+    vars:
+      rulepriority: "1"
+
+  # Extract the simple events
+  - import: org/clulab/asist/grammars/actions.yml
+    vars:
+      rulepriority: "5"
+
+  # Extract the internal simple events
+  - import: org/clulab/asist/internal/internal_actions.yml
+    vars:
+      rulepriority: "5"
+  # Extract the planning terms
+  - import: org/clulab/asist/grammars/planning.yml
+    vars:
+      rulepriority: "5"
+
+
+  # Extract the complex events
+  - import: org/clulab/asist/grammars/complex_actions.yml
+    vars:
+      rulepriority: "7+"
+
+  # Extract commitments
+  - import: org/clulab/asist/grammars/commitment.yml
+    vars:
+      rulepriority: "7+"
+
+  # Extract questions
+  - import: org/clulab/asist/grammars/questions.yml
+    vars:
+      rulepriority: "3+"
+
